adjust page.json format, add content replacement to assembler
`page.json`'s `content` node shall now be a dictionary of keys with
simple `{type, value}` dictionaries.  For `file` values, only the
basename is specified.  With this commit, the assembler contains support
for parsing this command for `string` and `file` source types.

This commit fixes the buildsystem such that it tracks rendered
prerequisites correctly.